Added: ParkingData, ParkingListData in entities. Parking, Parkings in server. (Parking has Many-To-One connection with ParkingLot and One-To-One connection with ParkingOrder)
Added:
When Customer order, after we build his order object, we search for free parking space in his ParkingLot, and we then update this parking space to points to this order and to be occupied.

Added:
Kiosk screen. customer can choose which parking lot he is in and take in or out his car. (Added a method that find the customer order (if he has) and also compute the late time if he is late)

30/01: fixed Kiosk class